pyproject.toml
[project]
name = "newsletter-agent"
version = "0.1.0"
description = "6-Agent Newsletter Generator powered by Google ADK × Qubrid AI × Composio"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.40.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "google-adk>=0.5.0",
    "litellm>=1.80.0",
    "composio-core>=0.7.0",
    "openai>=1.50.0",
    "ddgs>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false
